+++ src/main/java/com/scottlogic/compliance/AbstractChecker.java
@@ -3,6 +3,7 @@ package com.scottlogic.compliance;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.ConfigEvent;
+import com.scottlogic.compliance.event.*;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.config.ConfigClient;
 import software.amazon.awssdk.services.config.model.Evaluation;
@@ -18,25 +19,22 @@ import java.util.stream.Collectors;
 import static java.util.Objects.isNull;
 
 
-public abstract class AbstractChecker implements RequestHandler<ConfigEvent, Void>, Checkable {
+public abstract class AbstractChecker implements RequestHandler<ConfigEvent, Void>, Checkable, CheckableVisitor {
     private SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd_HH:mm:ss");
     private Context context;
 
     @Override
     public List<ComplianceResult> getComplianceCheck(ComplianceEvent event) {
-        if (event instanceof EventParser.ChangeEvent) {
-            return getComplianceCheck((EventParser.ChangeEvent) event);
-        } else if (event instanceof EventParser.PeriodicEvent) {
-            return getComplianceCheck((EventParser.PeriodicEvent) event);
-        }
-        throw new UnsupportedOperationException("BANG");
+        return event.accept(this);
     }
 
-    public List<ComplianceResult> getComplianceCheck(EventParser.ChangeEvent event) {
+    @Override
+    public List<ComplianceResult> getComplianceCheck(ComplianceChangeEvent event) {
         throw new UnsupportedOperationException("Config rule does not support Change Events");
     }
 
-    public List<ComplianceResult> getComplianceCheck(EventParser.PeriodicEvent event) {
+    @Override
+    public List<ComplianceResult> getComplianceCheck(CompliancePeriodicEvent event) {
         throw new UnsupportedOperationException("Config rule does not support Periodic Events");
     }
 
@@ -47,7 +45,7 @@ public abstract class AbstractChecker implements RequestHandler<ConfigEvent, Voi
         log("invokingEvent " + event.getInvokingEvent());
         log("ruleParameters " + event.getRuleParameters());
 
-        ComplianceEvent complianceEvent = EventParser.extract(event);
+        ComplianceEvent complianceEvent = ComplianceEventFactory.create(event);
 
         List<ComplianceResult> complianceResults = getComplianceCheck(complianceEvent);
