--- src/main/java/com/scottlogic/compliance/ComplianceEvent.java
@@ -1,3 +1,0 @@
-package com.scottlogic.compliance;
-
-public interface ComplianceEvent { }
