--- src/main/java/com/scottlogic/compliance/ParseException.java
@@ -1,22 +1,0 @@
-package com.scottlogic.compliance;
-
-public class ParseException extends RuntimeException {
-    public ParseException() {
-    }
-
-    public ParseException(String message) {
-        super(message);
-    }
-
-    public ParseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ParseException(Throwable cause) {
-        super(cause);
-    }
-
-    public ParseException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
