+++ src/main/java/com/scottlogic/compliance/event/ComplianceEvent.java
@@ -1,0 +1,9 @@
+package com.scottlogic.compliance.event;
+
+import com.scottlogic.compliance.ComplianceResult;
+
+import java.util.List;
+
+public interface ComplianceEvent {
+    List<ComplianceResult> accept (CheckableVisitor visitor);
+}
